mc_alpha_7
Biggest update so far. Spent a lot of time rethinking what will be
valuable to have in a prototype. This version gives the playtester cues
of what to do to get through the level. All objects now have scripts on
them with functions that trigger things (i.e. show a particular message
above this object on trigger enter and hide it on trigger exit). Also
player doesn't set values of other gameobjects, instead it calls a
method on the desired object's script. Feeling good about the
prototype. I also worked on making some art in illustrator, but
Margaret said that assets won't play nicely with github. For these
edits, I chose to just change colors when things happen.