SS_Alpha_22
Added the code to the elevator. Still having the same problem where the
gizmo can only move in one direction. For some reaon the logic trips up
and can never get to the right scenario where it can move in the
opposite direction. I plan to go talk to margret tomorrow if time
allows.